+++ cppad/local/parameter_op.hpp
@@ -63,7 +63,7 @@ is the zero order Taylor coefficient corresponding to z.
 \par Checked Assertions where op is the unary operator with one result:
 \li NumArg(op) == 1
 \li NumRes(op) == 1
-\li \a arg[0] < num_par 
+\li \a size_t(arg[0]) < num_par 
 \li \a 0 < \a nc_taylor
 */
 template <class Base>
@@ -78,7 +78,7 @@ inline void forward_par_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(ParOp) == 1 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(ParOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < num_par );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < num_par );
 	CPPAD_ASSERT_UNKNOWN( 0 < nc_taylor );
 
 	Base* z = taylor + i_z * nc_taylor;
