+++ cppad/local/sub_op.hpp
@@ -46,8 +46,8 @@ inline void forward_subvv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 
 	// Taylor coefficients corresponding to arguments and result
@@ -83,8 +83,8 @@ inline void forward_subvv_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 
 	// Taylor coefficients corresponding to arguments and result
 	Base* x = taylor + arg[0] * nc_taylor;
@@ -122,8 +122,8 @@ inline void reverse_subvv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
@@ -167,7 +167,7 @@ inline void forward_subpv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubpvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubpvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 
 	// Taylor coefficients corresponding to arguments and result
@@ -209,7 +209,7 @@ inline void forward_subpv_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubpvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubpvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 
 	// Paraemter value
 	Base x = parameter[ arg[0] ];
@@ -248,7 +248,7 @@ inline void reverse_subpv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
@@ -290,7 +290,7 @@ inline void forward_subvp_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvpOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvpOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 
 	// Taylor coefficients corresponding to arguments and result
@@ -334,7 +334,7 @@ inline void forward_subvp_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvpOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvpOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
 
 	// Parameter value
 	Base y = parameter[ arg[1] ];
@@ -373,7 +373,7 @@ inline void reverse_subvp_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(SubvpOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(SubvpOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
